Reorganized Utility code by service
