+++ src/CareTogether.Core/Resources/Accounts/AccountsResource.cs
@@ -1,4 +1,4 @@
-﻿using CareTogether.Resources.Storage;
+﻿using CareTogether.Utilities.ObjectStore;
 using System;
 using System.Collections.Immutable;
 using System.Linq;
