+++ src/CareTogether.Core/Resources/Policies/PoliciesResource.cs
@@ -1,4 +1,4 @@
-﻿using CareTogether.Resources.Storage;
+﻿using CareTogether.Utilities.ObjectStore;
 using System;
 using System.Collections.Immutable;
 using System.Threading.Tasks;
