+++ test/CareTogether.Core.Test/MemoryMultitenantObjectStore.cs
@@ -1,4 +1,4 @@
-﻿using CareTogether.Resources.Storage;
+﻿using CareTogether.Utilities.ObjectStore;
 using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
